+++ src/mkvtoolnix-gui/util/sleep_inhibitor_p.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include "common/common_pch.h"
+
+#include "common/debugging.h"
+
+namespace mtx { namespace gui { namespace Util {
+
+class BasicSleepInhibitor;
+
+class BasicSleepInhibitorPrivate {
+public:
+  std::vector<std::shared_ptr<BasicSleepInhibitor>> m_inhibitors;
+  static debugging_option_c ms_debug;
+};
+
+}}}